switch to authzed/cel-go
cel-go introduced breaking changes in v1.17, which SpiceDB now depends
on. But cel-go is a common dependency in several go projects, most
notably Kubernetes (which is still on 0.16.0)

Migrating to a fork allows SpiceDB to avoid conflicts with other
dependencies that use cel-go.

This should be a temporary change unless cel-go continues to make
breaking API changes.