makes PaginatedIterator.Close() idempotent
There are scenarios where a paginated iterator may
fail to fetch a page from the database after previous
successes. This would lead to a runtime panic
because the delegate datastore.RelationshipIterator could
be nil due to an error in startNewBatch(), and any call
to paginatedIterator.Close() would attempt
to call Close() on a nil delegate.