singleflight dispatch middleware
adds a new singleflight gRPC middleware capable of
deduping DispatchCheckRequest (for now, more to come
later)

The API hash computation logic was refactored so it could be
used outside of the services/v1 package

is uses janos/singleflight instead of x/sync/singleflight,
because the later does not propagate cancelation, and
we want to abort any expensive operation (especially
database operations) as soon as possible.

janos/single will let the handler run until the last
singleflighted caller is canceled. In exchange it has
to spin an extra goroutine compared to x/sync/singleflight.