singleflight dispatch middleware
adds a new singleflight `dispatch.Dispatcher` capable of
deduping `DispatchCheckRequest` (for now, more to come later).

is uses `janos/singleflight` instead of `x/sync/singleflight`,
because the latter does not propagate cancelation, and
we want to abort any expensive operation (especially database operations)
as soon as possible. janos/single will let the handler run until the
last single-flighted caller is canceled. In exchange,
it has to spin an extra goroutine compared to x/sync/singleflight.

I previously tried to implement it as a gRPC Dispatcher API middleware,
but that made things worse because we bypassed the caching dispatcher,
which happens at a different level. Adding it at the dispatcher level
proved to work as expected