log payloads and duration as an integer
in order to better troubleshoot certain situations,
it would be useful to load request and response payloads.

Also in certain log aggregators it's not possible to convert
duration fields that are represented as a string, and
this is useful in order to filter requests by a specific
duration threshold. This commit changes the duration function
to log it as an integer