Fix the caveat expr limit to be processed by our code
Right now, as a hack to ensure caveat error messages have the correct line and col information, we replace the preceding schema with newlines. However, the CEL parser has a default limit on the size of the expression it can be given, which was causing the error. With this change, we turn off the CEL check and instead check the *trimmed* expression ourselves for a limit.

Long term, we'll want to issue a change in CEL to allow updating of the positioning information instead of this hack

Adds tests and fixes #1637