fix inefficient singleflight deduplication with bloom filters
the previous approach caused an important regression
in deduplication rate because concurrent requests
that are part of the same requestID will trigger
the condition, but do not necessarily involve a loop.
We observed production systems that deduplicated
at a rate of 12% drop to 2%.

This approach uses a bloom filter to keep track of
what parts of the schema have been traversed. If there
is a false positive, the deduper will just dispatch, so
the loss in requests that could have been deduplicated
is, by default, 1%.

The number of elements in the bloom filter is by default
50, which comes from the default traversal depth, as we
know that no more than 50 different elements will be added,
since then the recursive call loop detection mechanism will
trigger.

The size of a 50 element bloom filter with 1% error rate
is 88 bytes.

The RequestID metadata was no longer needed
because the dispatcher becomes stateless
and the bloom filter is request specific. The
field was, however, marked as deprecated not to
break dispatch during a rollout of this change.

Tests were refactored to make them more readable,
and a new test was added to simulate a loop.