feat: Support fallback behavior when using old zedtokens
Before, a zedtoken from a SpiceDB on a different datastore could either fail with a cryptic error or simply have the wrong behavior.

After this change, the zedtoken is detected as belonging to a different datastore and a CLI option exists to either fail fully safe (switch to full consistency), to fail somewhat safe (switch to min latency) or to return an error (if requested).

To support this new behavior, a unique ID is attached to the datastore instance and a portion of its hash is including in the zedtoken. This makes zedtokens *slightly* larger.