reduces chunking allocations for wide relations
For most situations, when the number of elements to dispatch
is below the default 100 elements, the slice with elements
to dispatch was appropriately pre-allocated.

However, when dispatching large number of elements, the
silence will grow from 1 up to the number of chunks, causing
wasteful allocations in the critical path.

To simplify the logic, +1 is added to account for the situation
the number of elements is less than the chunk size. For values
above that it will cause one excess slice entry allocation, but seems
like a reasonable tradeoff w.r.t extra annoying code to handle it.