enables exemplars prometheus support
enabling exemplars makes possible to correlate
traces with a metric. Clients scraping will be
able to do content negotiation and request
OpenMetrics format.

This change is not innocuous: prometheus defaults
to scraping with OpenMetrics format, and as a result
the "le" labels of histograms will have a trailing zero.
From the prometheus/client_golang docs

// If true, the experimental OpenMetrics encoding is added to the
// possible options during content negotiation. Note that Prometheus
// 2.5.0+ will negotiate OpenMetrics as first priority. OpenMetrics is
// the only way to transmit exemplars. However, the move to OpenMetrics
// is not completely transparent. Most notably, the values of "quantile"
// labels of Summaries and "le" labels of Histograms are formatted with
// a trailing ".0" if they would otherwise look like integer numbers
// (which changes the identity of the resulting series on the Prometheus
// server).