CheckBulkPermissions
Adds the 'PermissionsService/CheckBulkPermissions' API method.
This promotes the existing 'ExperimentalService/BulkCheckPermission'
API out of the experimental service, but does not remove the old
method to ensure backwards compatibility.

To reduce code duplication, this refactors the bulk permission
checking logic into a shared component that can be used by both
the old and new methods. The old method performs protobuf message
conversion on the input, passes the inputs to the shared method,
and then converts the output.