report GC stats even in the event of a GC worker error
When customers seed their SpiceDB clusters and
use idempotent TOUCH operations, large amounts of garbage
may be generated that all will expire around the same time.

This makes the GC workers take longer to process, to the point
they may exceed the deadline. In these scenarios the code is
not reporting metrics, which hides how much garbage was being collected.

This changes the code so that GC metrics are issued, even in the event
of an error.