Use normal string instead of sql.NullString for caveatname when bulk importing tuples. Normal Writes also use an empty string. Touch operations with caveat updates wont work when the values are null instead of "".
Fixed problems with `<>` operator that will always return false (unknown)when `null` value is involved. `IS DISTINCT FROM` works the same way but treats `null` as a value and thus will work correctly in the case the value in the column is actually `null`.

Fixed usage of wrong column for caveat name.
Added test to check that update operations after a bulk load are working correctly.

Use correct formating of column name.

Reuse existing revisions.