enriches postgres revisions with txID and timestamp
consumers of the Datastore Watch API may want to access
the postgres transaction of a revision, and
the recorded timestamp.

Since we don't want to load it in every call sites for
performance reasons, a revision may be in a partial
state, thus it should be retrieved via the accessor methods.