issue a checkpoint when head revision moved outside an application transaction
it's possible for `pg_current_snapshot()` (used for HeadRevision) in PG
to return a new xmin:xmax outside of application transactions. For
example, running `ANALYZE;` could bump it.

In this situation, we need to emit a Checkpoint, in case clients
triggered a call to Watch API based on a different HeadRevision
from what they had cached locally (e.g. relevant to compute
change deltas between revisions).

The proposal is to compute head revision in the same transaction
where we compute new revisions, but only if checkpoints were requested
as an optimization. If the transaction determines there are no new
SpiceDB transactions, we return the compute head revision, if and only
if checkpoints were requested. This Watch API poll loop into 2 queries
instead of 1, which can add an extra load to the database, but at
least it only happens when checkpoints are being requested.