adds index to support efficient querying of Watch API
we identified Watch API was causing a full table scan on
large postgres-backed SpiceDB, overloading the database.

This commit adds an index that reduces the execution time
of relationship changes to be emitted via Watch API by
more than three orders of magnitude on tables with several
gigabytes worth of data.