make optimized revision resilient to out-of-band snapshot changes
this commit changes how postgres optimized revisions are
done to be able to quantize revisions event when
no application writes are happening and something
else in the postgres instance is causing the current snapshot
to move (e.g. VACUUM or non-spicedb workloads).

This commit is put in place to address the side-effects of
commit 3f22176ac3b5fc53ae959975c205b8b36513a826, where
we made PG watch API generate artificial new transaction
IDs as a checkpoint heart-beat. The fix in this commit
mitigates the impact to cache thrashing it would have caused.

It also fixes cache thrashing caused by sharing a postgres
database between multiple SpiceDB instances, or sharing
with other unrelated workloads.

The Query does the same thing as before, but now
COALESCES between the optimized revision selected, and the
latest revision available in the datastore. This means
that as time passes, the same last revision will always
be selected.

The LEFT JOIN is added so that COALESCE does not fail
with an empty result from the "optimized" CTE - it needs
a null result.