fixes problem with concurrent snapshots in optimized revision
this fixes the failing test-case in the previous commit,
demonstrating that the new query to determine
the optimized revision wasn't able to correctly
compute the revision if concurrent snapshots existed
in the same quantization window.

This fix changes the query so:
- if an optimized revision exists, it is returned
- if it does not, then all the revisions
  from the last quantization window are returned

Now the optimizedRevisionFunc needs to handle both scenarios:
- if an optimized revision was found, one row will be returned,
  and the revision will be returned as is
- if an optimized revision was not found, between we will get
  one or more rows. In either case, we will create a synthetic
  revision that captures all the XIDs in the quantization window.

The tradeoff here is multiple rows are returned. This should only
happen when no writes take place and the quantiation window is past.
In this case, it seems reasonable to tradeoff returning
multiple transaction rows for a deterministic optimized revision
that improves cache hit ratios.