postgres: optimize CheckRevision query
the condition in postgres' CheckRevision datastore
could lead to slow queries for very large
transactions table. This was caused by the
`timestamp = (SELECT MAX(timestamp) FROM relation_tuple_transaction)`
portion of the query, which led to filtering rows which
could have been discarded in the first place by the other expression.

This query was contributed by one of our customers.