fix spanner metrics regression
recent changes mean client-side spanner metrics
are no longer exposed the prometheus scrape endpoint.

The SDK has deprecated OpenCensus support, which was
wired into the prometheus endpoint, in favour of OpenTelemetry.
This was reflected in a recent SpiceDB change, which made the otel
approach the default, and as a consequence all prometheus metrics
disappeared.

This commit adds the prometheus exporter as a OpenTelemetry MetricsResolver,
to add some the metrics back. The naming convention has changed, and I'm not aware
of a way around that.

Also client-side latency metrics are not supported via the opentelemetry integration.
See https://cloud.google.com/spanner/docs/capture-custom-metrics-opentelemetry#capture-client-round-trip-latency
Instead they now have to be manually instrumented.