do not count closed conns due to cancellation toward health check
we noticed that when clients called APIs that propagate
cancellation error like WriteRelationships, it could cause
the write connection pool to get starved. This is because
the RetryPool considers that an error counts toward
the health check when it has caused a connection to be
closed.

This commit changes the code so this is ignored when
the connection is closed as part of context cancelation.
This should prevent marking as unhealthy a node that
is observing connections getting closed caused by the client
cancelation / deadlines.