don't acquire a new conn if there are no more retries
`RetryPool` will attempt to acquire a new connection
from a healthy node if the previous error was
considered resettable or the connection was closed.

This is unnecessary if a request is not going to retry,
like for example when `CtxDisableRetries` is in the context,
or when it is the last retry attempt. This adds
unnecessary pressure to the connection pool.

It would also cause requests that do not accept retries
to take longer to return, specially if there is already
contention in the connection pool.