Postgres: configure migration driver to support PgBouncer
In https://github.com/authzed/spicedb/pull/1994/commits/dcc797274edc59eb57222eefc0227442bb7abee2
we adjusted the postgres datastore handling of connections to use `exec` mode by default.
This fixes the compatibility issues SpiceDB had with PgBouncer.

However, that fix overlooked we had to do the same in the migrations driver, which
has its own logic to create the connections. We discovered this after
[a user report](https://github.com/orgs/authzed/discussions/2461) indicating
the failures emerged during migrations, but not on during normal operation.