*: replace exp/maps with maps
This mostly replaces usage of maps.Keys and
maps.Values with slices.Collect(maps.Keys/Values).

The one difference between exp/maps and maps
is that empty maps are never allocated with maps.

slices.AppendSeq(make([]K, 0, len(m)), maps.Keys(m))
is the the way to get the exact same behavior as
before, but I find it difficult to read and less
efficient because of the allocation.