Parse schema from schemafile property
Support for the `schemaFile` property in the validation file was added
in [#2206](https://github.com/authzed/spicedb/pull/2206) but it was
never fully integrated in to the bootstrap/import process.

This change adds support similar to the validator from the `zed` cli
which was added in [zed#478](https://github.com/authzed/zed/pull/478).

If both `schema` and `schemaFile` are specified, an error is thrown. If
only `schemaFile` is specified, the contents of the linked file are read
and passed to the compilation function. If `schema` is specified or
missing, this function behaves as it always has.