fix: duplicate metrics registration in mysql datastore
When using a replica with the mysql datastore, we register the
same metrics again and again.

This commit ignores AlreadyRegisteredError when registering our
custom metrics and adds a `db_name` label similar to the postgres
instrumentation.

Additionally, it updates the tests for multi instance creation in a way that we
don't create 2 connections with the same index. Otherwise we would
get another registration error when registering the sql collector `sqlstats`.

fixes: https://github.com/authzed/spicedb/issues/2518

Signed-off-by: leonnicolas <leonloechner@gmx.de>

Override the default Prometheus registry in tests again

Some tests were failing when previous test cases registered their
metrics with the default Prometheus registry.

Signed-off-by: leonnicolas <leonloechner@gmx.de>