chore: remove deprecated otel interceptors
This commit removes the usage of deprecated otel go SDK
gRPC interceptors. Those were superfluous since we
were already making use of the new StatsHandler
installed in the gRPC server.

I verified that traces are being correctly collected
via otel, and that the TraceID shows up in the logs.
Also tested the grpc Gateway, but did not verify
the Dispatch API.

I also identified this was missing in the testserver,
so I added it, along with missing request logs.