fix(api): breakage of gRPC Proposal A6: Retry policies
See https://github.com/grpc/proposal/blob/master/A6-client-retries.md#when-retries-are-valid

As I was trying to implement an example client
using retry policies, I found they didn't work.
This was due to the fact SpiceDB response
violated the expectations from that spec, notably:

> In certain cases it is not valid to retry an RPC. These cases occur when the RPC has been committed, and thus it does not make sense to perform the retry.
> The reasoning behind the first scenario is
that the Response-Headers include initial metadata
from the server. The metadata (or its absence)
it is transmitted to the client application.
This may fundamentally change the state of the
client, so we cannot safely retry if
a failure occurs later in the RPC’s life.

The RequestID middleware was sending the requestID as a
metadata header, which causes the stream to send a
frame just to send the headers before sending the actual
response. This in turn causes the stream to become committed,
which as per A6 spec causes retry policies to be disabled.

The TL;DR is: use trailers for metadata unless there is
an explicit reason to use metadata headers, otherwise
retry policies will become disabled.