feat: add Memory Protection Middleware
The commit introduces a new memory protection middleware to help prevent out-of-memory conditions in SpiceDB by implementing admission control based on current memory usage.

This is not a perfect solution (doesn't prevent non-traffic-related sources of OOM) and is meant to support other future improvements to resource sharing in a single SpiceDB node.

The middleware is installed both in the main api and in dispatch, but at
different thresholds. Memory usage is polled in the background, and if
in-flight memory rises above the threshold, backpressure is placed on
incoming requests.

The API threshold is higher than the dispatch threshold to preserve
already admitted traffic as much as possible.