fix(pkg): isolates dependency with github.com/odigos-io/go-rtml
The introduction of "github.com/odigos-io/go-rtml" as
a dependency has caused many go toolchain operations to fail
because it now requires special flags to resolve a runtime symbol.

This is not ideal because it breaks the developer experience
for folks using IDEs with native Go support. Something as
simple as go test ./... no longer works without flags.

But most importantly: it becomes sticky to any go module importing
SpiceDB.

This commit isolates the memory protection usage by:
- adding a global variable that acts as the registry
- isolating the MemoryUsageProvider into its own package
- introducing a new NoopMemoryLimiter, and using it everywhere
  we explicitly used the go-rtml package
- moving tests into isolated packages
- initializing the global default provider in main.go, using a function
  controlled by build tags. Without memoryprotection build tag, defaults
  to using the noop implementation.
- moving all tests in the main package to make sure one can run them
  without the need of passing the special flags.

Consequence:
- now all tests run without special flags, except for those where
  we explicitly test for this behavior.
- go modules importing SpiceDB do won't require the flags.
- folks can run SpiceDB in their IDEs quickly without tinkering
  with toolchain flags